Module Ft_turing.Utils

exception Error

val error : string -> 'a

error msg used for all matters of interruption of the program thoughout execution

val err_invalid_type : bool -> Yojson.Basic.t -> string -> 'a

err_invalid_type std t field is called whenever an error of typing occurs during parsing of json file

val wrap_error : 'a -> 'a

wrap_error f

val assert_transition_ok : (string * 'a, string * 'b * 'c) Stdlib.Hashtbl.t -> (string * 'b * 'c) -> string -> 'a -> string list -> unit

assert_transition_ok transitbl t state_key read_key alphabet states asserts a given transition is coherent and complies with specifications of turing machine definition.